refactored some more file structure
